+++ engine/src/WindowGLFW.inl
@@ -15,7 +15,7 @@ constexpr inline const GLFWwindow *const Engine::Core::Renderering::WindowGLFW::
     return m_window;
 }
 
-inline void Engine::Core::Renderering::WindowGLFW::getSize(size_t &width, size_t &height) const noexcept
+inline void Engine::Core::Renderering::WindowGLFW::getSize(unsigned short &width, unsigned short &height) const noexcept
 {
     glfwGetWindowSize(m_window, reinterpret_cast<int *>(&width), reinterpret_cast<int *>(&height));
 }
